make 2.9.1 build but Mutabor crashes still at runtime in an infinite recursion :-(

+++ mu32/routing/gmn/GSP.cpp
@@ -636,7 +636,7 @@ int DoParse()
 					              Brackets[Brackets.Len()-1]);
 
 					if ( Brackets[Brackets.Len()-1] != i-1 )
-						return DoError(2 + Brackets[BracketDeep-1]/2);
+						return DoError(2 + (int) Brackets[BracketDeep-1]/2);
 
 					Brackets = Brackets.Left(Brackets.Len()-1);
